requirements.txt
# Bible Podcaster - Минимальные зависимости

# Core
requests
python-dotenv
pydantic
pydantic-settings

# AI/Text Processing
openai
anthropic

# Audio Generation
elevenlabs
pydub

# Image Processing
Pillow
opencv-python

# Video Processing
moviepy
imageio
imageio-ffmpeg

# YouTube API
google-api-python-client
google-auth
google-auth-oauthlib

# Logging
structlog
rich

# Testing
pytest 